+++ nexus/client.go
@@ -413,92 +413,6 @@ func (c *Client) sendGetOperationRequest(ctx context.Context, request *http.Requ
 	}
 }
 
-// NewCompletionHTTPRequest creates an HTTP request deliver an operation completion to a given URL.
-func NewCompletionHTTPRequest(ctx context.Context, url string, completion OperationCompletion) (*http.Request, error) {
-	httpReq, err := http.NewRequestWithContext(ctx, "POST", url, nil)
-	if err != nil {
-		return nil, err
-	}
-	if err := completion.applyToHTTPRequest(httpReq); err != nil {
-		return nil, err
-	}
-
-	httpReq.Header.Set(headerUserAgent, userAgent)
-	return httpReq, nil
-}
-
-// OperationCompletion is input for [NewCompletionHTTPRequest].
-// It has two implementations: [OperationCompletionSuccessful] and [OperationCompletionUnsuccessful].
-type OperationCompletion interface {
-	applyToHTTPRequest(*http.Request) error
-}
-
-// OperationCompletionSuccessful is input for [NewCompletionHTTPRequest], used to deliver successful operation results.
-type OperationCompletionSuccessful struct {
-	// Header to send in the completion request.
-	Header http.Header
-	// Body to send in the completion HTTP request.
-	// If it implements `io.Closer` it will automatically be closed by the client.
-	Body io.Reader
-}
-
-// NewOperationCompletionSuccessful constructs an [OperationCompletionSuccessful] from a JSONable value.
-// Marshals the provided value to JSON using [json.Marshal] and sets the proper Content-Type header.
-func NewOperationCompletionSuccessful(v any) (*OperationCompletionSuccessful, error) {
-	b, err := json.Marshal(v)
-	if err != nil {
-		return nil, err
-	}
-
-	header := make(http.Header, 1)
-	header.Set(headerContentType, contentTypeJSON)
-
-	return &OperationCompletionSuccessful{
-		Header: header,
-		Body:   bytes.NewReader(b),
-	}, nil
-}
-
-func (c *OperationCompletionSuccessful) applyToHTTPRequest(request *http.Request) error {
-	if c.Header != nil {
-		request.Header = c.Header.Clone()
-	}
-	request.Header.Set(headerOperationState, string(OperationStateSucceeded))
-	if closer, ok := c.Body.(io.ReadCloser); ok {
-		request.Body = closer
-	} else {
-		request.Body = io.NopCloser(c.Body)
-	}
-	return nil
-}
-
-// OperationCompletionUnsuccessful is input for [NewCompletionHTTPRequest], used to deliver unsuccessful operation
-// results.
-type OperationCompletionUnsuccessful struct {
-	// Header to send in the completion request.
-	Header http.Header
-	// State of the operation, should be failed or canceled.
-	State OperationState
-	// Failure object to send with the completion.
-	Failure *Failure
-}
-
-func (c *OperationCompletionUnsuccessful) applyToHTTPRequest(request *http.Request) error {
-	if c.Header != nil {
-		request.Header = c.Header.Clone()
-	}
-	request.Header.Set(headerOperationState, string(c.State))
-	request.Header.Set(headerContentType, contentTypeJSON)
-
-	b, err := json.Marshal(c.Failure)
-	if err != nil {
-		return err
-	}
-
-	request.Body = io.NopCloser(bytes.NewReader(b))
-	return nil
-}
-
 // readAndReplaceBody reads the response body in its entirety and closes it, and then replaces the original response
 // body with an in-memory buffer.
 // The body is replaced even when there was an error reading the entire body.
